backend/response_generation/pyproject.toml
[project]
name = "response-generation"
version = "0.1.0"
description = "Agent-based response generation with OpenRouter integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "pydantic>=2.10.3",
    "python-dotenv>=1.0.1",
    "openai>=1.57.4",
    "uvicorn>=0.34.0",
    "pytest>=8.3.4",
    "httpx>=0.27.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
]